Fix bug when be using on a PC.
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk/classes@3232 06c8f799-c779-4783-8af7-179afedb67c8

+++ lejos/util/Delay.java
@@ -24,7 +24,7 @@ public class Delay {
             {
                 interrupted = true;
             }
-            period = end - (int)System.currentTimeMillis();
+            period = end - System.currentTimeMillis();
         } while (period > 0);
         if (interrupted)
             Thread.currentThread().interrupt();
